rename LIP12 streamflow
